+++ ui/litellm-dashboard/eslint-metrics.json
@@ -1,5 +1,5 @@
 {
-  "@typescript-eslint/no-explicit-any": 1978,
+  "@typescript-eslint/no-explicit-any": 1977,
   "complexity": 129,
   "local/no-large-inline-object-arg": 512,
   "local/no-long-condition-chain": 233,
